feature(web/src/app/examples/layours): Separating into components: list of users, user table and the options.
