+++ web/src/app/examples/layouts/table-example-layout.tsx
@@ -1,0 +1,115 @@
+import { useEffect, useState } from "react";
+
+import { getUsers } from "@/core/services/user-service";
+import User from "@/core/interfaces/user.interface";
+import { inter } from "@/ui/fonts";
+
+export default function TableExampleLayout() {
+  const [users, setUsers] = useState<User[]>([]);
+  const [loading, setLoading] = useState<boolean>(false);
+  const [error, setError] = useState<string | null>(null);
+
+  useEffect(() => {
+    const fetchUsers = async () => {
+      try {
+        setLoading(true);
+        const users = await getUsers();
+        console.log({ users });
+        setUsers(users);
+      } catch (error) {
+        setError("Error fetching users.");
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchUsers();
+  }, []);
+
+  return (
+    <div className="container my-5">
+      <div className="container">
+        <p
+          className={`my-4 uppercase text-3xl underline text-center font-bold tracking-wide`}
+        >
+          consuming apirest in springboot to view table
+        </p>
+      </div>
+
+      <div className="overflow-x-auto shadow-md sm:rounded-lg">
+        <table className="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400">
+          <thead className="text-xs text-white uppercase bg-gray-500 dark:bg-gray-700 dark:text-gray-400">
+            <tr>
+              <th scope="col" className="px-6 py-3">
+                Name user
+              </th>
+              <th scope="col" className="px-6 py-3">
+                id
+              </th>
+              <th scope="col" className="px-6 py-3">
+                email
+              </th>
+              <th scope="col" className="px-6 py-3">
+                birthdate
+              </th>
+              <th scope="col" className="px-6 py-3">
+                <span className="sr-only">Edit</span>
+              </th>
+            </tr>
+          </thead>
+          <tbody className="bg-slate-200">
+            {users.length === 0 ? (
+              <tr>
+                <td colSpan={5} className="px-6 py-4 text-center">
+                  No existen usuarios en la base de datos
+                </td>
+              </tr>
+            ) : (
+              users.map((user, index) => (
+                <tr
+                  key={index}
+                  className="bg-white border-b dark:bg-gray-800 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-600"
+                >
+                  <th
+                    scope="row"
+                    className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white"
+                  >
+                    {user.name}
+                  </th>
+                  <td className="px-6 py-4">{user.id}</td>
+                  <td className="px-6 py-4">
+                    <b>{user.email}</b>
+                  </td>
+                  <td className="px-6 py-4 ">
+                    <b>{new Date(user.birthdate).toLocaleDateString()}</b>
+                  </td>
+                  <td className="px-6 py-4 text-right">
+                    <button
+                      type="button"
+                      className="text-white bg-[#3b5998] hover:bg-[#3b5998]/90 focus:ring-4 focus:outline-none focus:ring-[#3b5998]/50 font-medium rounded-lg text-sm px-5 py-2.5 text-center inline-flex items-center dark:focus:ring-[#3b5998]/55 me-2 mb-2"
+                    >
+                      <svg
+                        className="w-4 h-4 me-2"
+                        aria-hidden="true"
+                        xmlns="http://www.w3.org/2000/svg"
+                        fill="currentColor"
+                        viewBox="0 0 8 19"
+                      >
+                        <path
+                          fillRule="evenodd"
+                          d="M6.135 3H8V0H6.135a4.147 4.147 0 0 0-4.142 4.142V6H0v3h2v9.938h3V9h2.021l.592-3H5V3.591A.6.6 0 0 1 5.592 3h.543Z"
+                          clipRule="evenodd"
+                        />
+                      </svg>
+                      Edit profile
+                    </button>
+                  </td>
+                </tr>
+              ))
+            )}
+          </tbody>
+        </table>
+      </div>
+    </div>
+  );
+}
